pyproject.toml
[project]
name = "dbt-graphql"
version = "0.1.0"
description = "Convert dbt artifacts to GraphQL schema and serve a SQL-backed GraphQL API"
readme = "README.md"
requires-python = ">=3.13,<3.14"
license = "MIT"
dependencies = [
    "dbt-artifacts-parser>=0.13.0",
    "dbt-colibri>=0.3.0",
    "pyyaml>=6.0.2",
    "sqlalchemy[asyncio]>=2.0.49",
]

[project.optional-dependencies]
duckdb = ["duckdb-engine>=0.17"]
sqlite = ["aiosqlite>=0.22.1"]
postgres = ["asyncpg>=0.31"]
mysql = ["aiomysql>=0.3.2"]
api = ["ariadne>=1.0.1", "fastapi>=0.136.0", "granian>=2.7.3"]
mcp = ["fastmcp>=3.2.4"]

[project.scripts]
dbt-graphql = "dbt_graphql.cli:main"

[dependency-groups]
test = ["pytest", "pytest-cov", "pytest-asyncio", "pytest-docker", "httpx"]
lint = ["ruff", "ty"]

[build-system]
requires = ["uv_build>=0.11.6,<0.12"]
build-backend = "uv_build"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
omit = ["tests/*", "**/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
